+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Changes
 
+- yofi.config is deprecated.
+
 ## Fixes
 
 # 0.2.2 - 2024-03-10